GUAC-1100: Make ObjectPermission concrete. Remove type-specific implementations.
