+++ guacamole-ext/src/main/java/org/glyptodon/guacamole/net/auth/permission/ObjectPermission.java
@@ -28,9 +28,10 @@ package org.glyptodon.guacamole.net.auth.permission;
  * whole.
  *
  * @author Michael Jumper
- * @param <T> The type of identifier used by the object this permission affects.
+ * @param <IdentifierType>
+ *     The type of identifier used by the object this permission affects.
  */
-public interface ObjectPermission<T> extends Permission<ObjectPermission.Type> {
+public class ObjectPermission<IdentifierType> implements Permission<ObjectPermission.Type> {
 
     /**
      * Specific types of object-level permissions. Each permission type is
@@ -61,12 +62,80 @@ public interface ObjectPermission<T> extends Permission<ObjectPermission.Type> {
     }
 
     /**
+     * The identifier of the GuacamoleConfiguration associated with the
+     * operation affected by this permission.
+     */
+    private final IdentifierType identifier;
+
+    /**
+     * The type of operation affected by this permission.
+     */
+    private final Type type;
+
+    /**
+     * Creates a new ObjectPermission having the given type and identifier.
+     * The identifier must be the unique identifier assigned to the object
+     * associated with this permission by the AuthenticationProvider in use.
+     *
+     * @param type
+     *     The type of operation affected by this permission.
+     *
+     * @param identifier
+     *     The identifier of the object associated with the operation affected
+     *     by this permission.
+     */
+    public ObjectPermission(Type type, IdentifierType identifier) {
+
+        this.identifier = identifier;
+        this.type = type;
+
+    }
+
+   /**
      * Returns the identifier of the specific object affected by this
      * permission.
      *
      * @return The identifier of the specific object affected by this
      *         permission.
      */
-    public T getObjectIdentifier();
+    public IdentifierType getObjectIdentifier() {
+        return identifier;
+    }
+
+    @Override
+    public Type getType() {
+        return type;
+    }
+
+    @Override
+    public int hashCode() {
+        int hash = 5;
+        if (identifier != null) hash = 47 * hash + identifier.hashCode();
+        if (type != null)       hash = 47 * hash + type.hashCode();
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+
+        // Not equal if null or wrong type
+        if (obj == null) return false;
+        if (getClass() != obj.getClass()) return false;
+
+        final ObjectPermission other = (ObjectPermission) obj;
+
+        // Not equal if different type
+        if (this.type != other.type)
+            return false;
+
+        // If null identifier, equality depends on whether other identifier
+        // is null
+        if (identifier == null)
+            return other.identifier == null;
+
+        // Otherwise, equality depends entirely on identifier
+        return identifier.equals(other.identifier);
+
+    }
 
 }
